DjangoFitnessApp/fitnessapp/users/views/logout_view.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.exceptions import AuthenticationFailed
from ..models import User
from ..serializers import UserSerializer
import jwt, datetime

class LogoutView(APIView):
    def post(self, request):
        response = Response()
        response.delete_cookie('jwt')
        response.data = {
            'message': 'success'
        }
        return response